# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Social Network API",
        "description": "The public API from the Social Network team.",
        "contact": {
            "name": "the GTI525 Social Network Team."
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "https://gti525-social-network.herokuapp.com/",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/user/tickets": {
            "get": {
                "tags": [
                    "tickets"
                ],
                "summary": "List all tickets of an authenticated user.",
                "operationId": "getTickets",
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "schema": {
                            "$ref": "#/definitions/Ticket"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "tickets"
                ],
                "summary": "Add one or multiple tickets to authenticated user.",
                "operationId": "addTicket",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "body",
                        "description": "A ticket to add to a user.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TicketPost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Resource created.",
                        "schema": {
                            "$ref": "#/definitions/Ticket"
                        }
                    },
                    "404": {
                        "description": "User not found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/ticket/{ticket_ID}": {
            "get": {
                "tags": [
                    "tickets"
                ],
                "summary": "Show a specific ticket belonging to a specific authenticated user.",
                "operationId": "getTicket",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "ticket_ID",
                        "in": "path",
                        "description": "Target ticket.",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "schema": {
                            "$ref": "#/definitions/Ticket"
                        }
                    },
                    "404": {
                        "description": "Ticket not found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/ticket/{unique_ticket_id}/owner/{new_user_id}": {
            "patch": {
                "tags": [
                    "tickets"
                ],
                "summary": "Change owner of a ticket_id to the selected user_id",
                "operationId": "transferTicket",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "unique_ticket_id",
                        "in": "path",
                        "description": "The unique ID of a ticket.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "new_user_id",
                        "in": "path",
                        "description": "The user ID to which we want to transfer ticket.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Ticket"
                            }
                        }
                    },
                    "403": {
                        "description": "Ticket does not belongs to you. Cannot transfer ticket.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "400": {
                        "description": "Ticket or user not found. Cannot transfer ticket.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/{friend_ID}/tickets": {
            "get": {
                "tags": [
                    "tickets"
                ],
                "summary": "Show the list of friend's tickets.",
                "operationId": "getFriendTicket",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "friend_ID",
                        "in": "path",
                        "description": "Target friend.",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Ticket"
                            }
                        }
                    },
                    "403": {
                        "description": "You are not friend with this person. You cannot retrieve its tickets.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Error": {
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "description": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "Ticket": {
            "required": [
                "id",
                "unique_id",
                "ownerFirstName",
                "ownerLastName",
                "user_id",
                "event",
                "artist",
                "price",
                "venue",
                "city",
                "image",
                "date",
                "description"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "unique_id": {
                    "type": "string"
                },
                "ownerFirstName": {
                    "type": "string"
                },
                "ownerLastName": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer",
                    "format": "int32"
                },
                "event": {
                    "type": "string"
                },
                "artist": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "format": "float"
                },
                "venue": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "date_event": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "TicketPost": {
            "required": [
                "id",
                "unique_id",
                "ownerFirstName",
                "ownerLastName",
                "user_id",
                "event",
                "artist",
                "price",
                "venue",
                "city",
                "image",
                "date",
                "description"
            ],
            "properties": {
                "unique_id": {
                    "type": "string"
                },
                "event": {
                    "type": "string"
                },
                "artist": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "format": "float"
                },
                "venue": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "date_event": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
...
    }
}